Move StringExtensions to portable library
