+++ Source/HelixToolkit.Wpf.Tests/HelixToolkit.Wpf.Tests.csproj
@@ -84,7 +84,6 @@
     <Compile Include="Exporters\ObjExporterTests.cs" />
     <Compile Include="Exporters\ExporterTests.cs" />
     <Compile Include="Exporters\X3DExporterTests.cs" />
-    <Compile Include="ExtensionMethods\StringExtensionsTests.cs" />
     <Compile Include="Geometry\BoundingSphereTests.cs" />
     <Compile Include="Geometry\CuttingEarsTriangulatorTests.cs" />
     <Compile Include="Geometry\LineSegmentTests.cs" />
